// app/common/enum.ts
export enum RESPONSE {
  SUCCESS,
  ERROR,
}
// ...
export enum ENUM_ENDPOINT_RPC {
  _NEMO_WALLET,
  _GALIXCITY,
  _MECHA_WARFARE,
  _9DNFT,
  _FLASHPOINT,
  _RICHWORK_FARM_FAMILY,
}

export enum TAB_BOTTOM_INO_BOX {
  OVERVIEW = "overview",
  RULE = "rule",
  VIDEO = "video",
}

export enum TYPE_GAME_MECHA {
  MEDAL = "paladin medal",
  WEAPON = "weapon",
  VALARION_R = "valarion r",
  VALARION_SR = "valarion sr",
  VALARION_SSR = "valarion ssr",
  VALARION_UR = "valarion ur",
}

export enum RESPONSE_CODE_RPC {
  _UNAUTHORIZED = 4100,
}